requirements.txt
PyQt5
pywin32
Pillow
Flask
requests 